feat: 투표 상세페이지 qa+컴포넌트 분리
