+++ src/features/vote/components/VoteItem/style.tsx
@@ -18,5 +18,8 @@ export const VsBox = styled.div`
 `;
 
 export const VoteImageWrapper = styled.div`
+  display: flex;
+  flex-direction: column;
+  align-items: center;
   width: 50%;
 `;
