+++ src/pages/Vote/VoteDetail/index.tsx
@@ -1,11 +1,10 @@
-import { MouseEvent } from 'react';
 import { useParams } from 'react-router-dom';
 import { useQuery } from '@tanstack/react-query';
 import { CommonDrawer, CommonIconButton, CommonText, DateText, Header } from '@/shared/components';
-import { useAuthCheck, useCustomToast, useDrawer } from '@/shared/hooks';
+import { useDrawer } from '@/shared/hooks';
 import { Body, Content, Footer, Span, Title } from './style';
 import { VoteOptionItem } from '@/features/vote/components';
-import { useCancelVote, useDeleteVote, useParticipationVote } from '@/features/vote/hooks';
+import { useDeleteVote } from '@/features/vote/hooks';
 import { voteQueryOption } from '@/features/vote/service';
 
 const VoteDetail = () => {
@@ -13,26 +12,7 @@ const VoteDetail = () => {
   const numberVoteId = Number(voteId);
   const { data: voteDetailData } = useQuery({ ...voteQueryOption.detail(numberVoteId) });
   const { mutate: DeleteVoteMutate } = useDeleteVote();
-  const { mutate: ParticipationVoteMutate } = useParticipationVote(numberVoteId);
-  const { mutate: CancelVoteMutate } = useCancelVote(numberVoteId);
-  const openToast = useCustomToast();
-  const isLogin = useAuthCheck();
   const { isOpen, onOpen, onClose } = useDrawer();
-
-  const toggleVoteParticipation = (e: MouseEvent<HTMLButtonElement>) => {
-    const { value } = e.currentTarget;
-    if (voteDetailData?.voteInfo.isVoting && isLogin) {
-      if (voteDetailData?.selectedItemId !== Number(value)) {
-        ParticipationVoteMutate({ voteId: numberVoteId, itemId: Number(value) });
-      } else {
-        CancelVoteMutate(numberVoteId);
-      }
-    } else if (!voteDetailData?.voteInfo.isVoting) {
-      openToast({ type: 'error', message: '종료된 투표에는 참여하실수 없습니다.' });
-    } else {
-      openToast({ type: 'error', message: '로그인이 필요한 서비스입니다.' });
-    }
-  };
 
   return (
     <>
@@ -49,15 +29,17 @@ const VoteDetail = () => {
         </CommonText>
         <Content>
           <VoteOptionItem
+            voteDetailData={voteDetailData!}
             itemInfo={voteDetailData?.item1Info}
             votes={voteDetailData?.voteInfo.item1Votes}
-            onClick={toggleVoteParticipation}
+            voteId={numberVoteId}
           />
           <Span>VS</Span>
           <VoteOptionItem
+            voteDetailData={voteDetailData!}
             itemInfo={voteDetailData?.item2Info}
             votes={voteDetailData?.voteInfo.item2Votes}
-            onClick={toggleVoteParticipation}
+            voteId={numberVoteId}
           />
         </Content>
         <Footer>
